requirements.txt
fastapi
uvicorn
pydantic_settings
openpyxl
pandas